pyproject.toml
[project]
name = "believers-eva-ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.1",
    "pydantic-ai>=1.7.0",
]
